requirements.txt
# Dependências do Extrator de Contratos Raízen
# Instalar com: pip install -r requirements.txt

# Extração de PDF
pdfplumber>=0.10.0

# Manipulação de Dados e Excel
pandas>=2.0.0
openpyxl>=3.0.0

# Configuração
PyYAML>=6.0

# OCR Fallback (para PDFs escaneados/imagem)
easyocr>=1.7.0

# Opcional - para desenvolvimento
pytest>=7.0.0
pytest-cov>=4.0.0

# Validação e Testes de Regressão
deepdiff>=7.0.0

# Automação com IA (Golden Set)
google-generativeai>=0.3.0
python-dotenv>=1.0.0
